+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 ### Bug fixes
 
 * [#4105](https://github.com/bbatsov/rubocop/issues/4105): Fix `Lint/IndentationWidth` when `Lint/EndAlignment` is configured with `start_of_line`. ([@brandonweiss][])
+* [#5453](https://github.com/bbatsov/rubocop/issues/5453): Fix erroneous downcase in `Performance/Casecmp` auto-correction. ([@walinga][])
 * [#5343](https://github.com/bbatsov/rubocop/issues/5343): Fix offense detection in `Style/TrailingMethodEndStatement`. ([@garettarrowood][])
 * [#5334](https://github.com/bbatsov/rubocop/issues/5334): Fix semicolon removal for `Style/TrailingBodyOnMethodDefinition` autocorrection. ([@garettarrowood][])
 * [#5350](https://github.com/bbatsov/rubocop/issues/5350): Fix `Metric/LineLength` false offenses for URLs in double quotes. ([@garettarrowood][])
@@ -3181,3 +3182,4 @@
 [@albertpaulp]: https://github.com/albertpaulp
 [@orgads]: https://github.com/orgads
 [@leklund]: https://github.com/leklund
+[@walinga]: https://github.com/walinga